Refactor to standard package layout
