+++ circuitron/__init__.py
@@ -1,0 +1,5 @@
+"""Circuitron package."""
+
+__version__ = "0.1.0"
+__author__ = "Circuitron Development Team"
+
