+++ circuitron/cli.py
@@ -1,11 +1,7 @@
-"""
-Main entry point for the Circuitron development system.
-Orchestrates the agent pipeline and handles command-line interface.
-"""
+"""Command line interface for Circuitron."""
 
 import asyncio
 from .config import setup_environment
-from circuitron.pipeline import pipeline, parse_args
 from .models import PartFinderOutput
 
 
@@ -13,11 +9,16 @@ async def run_circuitron(
     prompt: str, show_reasoning: bool = False, debug: bool = False
 ) -> PartFinderOutput:
     """Execute the Circuitron workflow using the full pipeline."""
+    from circuitron.pipeline import pipeline
+
     return await pipeline(prompt, show_reasoning=show_reasoning, debug=debug)
 
 
 def main():
     """Main entry point for the Circuitron system."""
+    setup_environment()
+    from circuitron.pipeline import pipeline, parse_args
+
     args = parse_args()
     prompt = args.prompt or input("Prompt: ")
     show_reasoning = args.reasoning
