--- development/__init__.py
@@ -1,8 +1,0 @@
-"""
-Circuitron Development Package
-
-A modular AI-driven PCB design system with multiple specialized agents.
-"""
-
-__version__ = "0.1.0"
-__author__ = "Circuitron Development Team"
